Refs #34
Missing categories were already being not shown for other tables related to categories/field ids but there was a bug in the system where the mapping went wrong when the category ids and field ids come with irregular ids . This is solved and mapping is now comign proper. Tested on the live site data exported.